+++ src/components/Admin/index.ts
@@ -1,5 +1,4 @@
 export * from './Group/GroupPage';
 export * from './ListPage';
-export * from './User/UserForm';
 export * from './User/UserList';
 export * from './User/UserPage';
