+++ src/components/UserProfile/UserForm.tsx
@@ -3,20 +3,20 @@ import React, { FC, useMemo } from 'react';
 import { Col, Form } from 'react-bootstrap';
 import { useTranslation } from 'react-i18next';
 import * as yup from 'yup';
-import { useTagsetsTemplateQuery } from '../../../generated/graphql';
-import { Tagset } from '../../../models/Profile';
-import { defaultUser, UserFromGenerated, UserModel } from '../../../models/User';
-import { TagsetTemplate } from '../../../types/graphql-schema';
-import countriesList from '../../../utils/countriesList.json';
-import { EditMode } from '../../../utils/editMode';
-import Button from '../../core/Button';
-import Loading from '../../core/Loading';
-import SearchDropdown from '../../core/SearchDropdown';
-import Section, { Header } from '../../core/Section';
-import EditableAvatar from '../../EditableAvatar';
-import { InputField } from '../Common/InputField';
-import { ReferenceSegment } from '../Common/ReferenceSegment';
-import TagsetSegment from '../Common/TagsetSegment';
+import { useTagsetsTemplateQuery } from '../../generated/graphql';
+import { Tagset } from '../../models/Profile';
+import { defaultUser, UserFromGenerated, UserModel } from '../../models/User';
+import { TagsetTemplate } from '../../types/graphql-schema';
+import countriesList from '../../utils/countriesList.json';
+import { EditMode } from '../../utils/editMode';
+import Button from '../core/Button';
+import Loading from '../core/Loading';
+import SearchDropdown from '../core/SearchDropdown';
+import Section, { Header } from '../core/Section';
+import EditableAvatar from '../EditableAvatar';
+import { InputField } from '../Admin/Common/InputField';
+import { ReferenceSegment } from '../Admin/Common/ReferenceSegment';
+import TagsetSegment from '../Admin/Common/TagsetSegment';
 
 interface UserProps {
   user?: UserModel;
